+++ python/cudf/cudf/core/dtypes.py
@@ -241,3 +241,7 @@ class DecimalDtype(ExtensionDtype):
     @classmethod
     def from_arrow(cls, typ):
         return cls(typ.precision, typ.scale)
+
+    @property
+    def itemsize(self):
+        return 8